Use loaders for fetching articles and learningpaths
